backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
cryptography==41.0.4
supabase==2.0.3
firebase-admin==6.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
email-validator==2.0.0
Flask-Mail==0.9.1
python-dateutil==2.8.2
urllib3==1.26.18

# Testing dependencies
pytest==7.4.0
pytest-flask==1.3.0
pytest-mock==3.11.1
pytest-cov==4.1.0
requests-mock==1.11.0
